+++ src/RoutesHostClient/UnavailableRoute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RoutesHostClient
+{
+	public class UnavailableRoute
+	{
+		public string Address { get; set; }
+		public DateTime ReleaseDate { get; set; }
+	}
+}