+++ src/RoutesHostServer/Models/ResolveResult.cs
@@ -7,6 +7,14 @@ namespace RoutesHostServer.Models
 {
 	public class ResolveResult
 	{
+		public ResolveResult()
+		{
+			AddressList = new List<string>();
+		}
+
+		[Obsolete("Use AddressList instead (Property removed in next version)", false)]
 		public string Address { get; set; }
+
+		public List<string> AddressList { get; set; }
 	}
 }